+++ src/modules/careplan/providers/aha/aha.careplan.service.ts
@@ -994,6 +994,10 @@ export class AhaCareplanService implements ICareplanService {
         }
     }
 
+    scheduleDailyHighRiskCareplan(): Promise<void> {
+        throw new Error("Method not implemented.");
+    }
+
     //#endregion
 
 }
